refactor: 메뉴 리스트 변경 함수 구현 및 별도 파일 분리
