+++ style.css
@@ -68,10 +68,12 @@ main {
     gap: 10px;
     flex: 1 1 45%;
     > img {
+      display: block;
       width: 200px;
       height: 150px;
       border: 2px solid #8f8f8f;
       border-radius: 4px;
+      object-fit: cover;
     }
     .menu_item {
       display: flex;
